sensor-node/requirements.txt
# SmartSense Sensor Node - Production Dependencies (Raspberry Pi)
# For deployment on Raspberry Pi with actual hardware

# MQTT Communication
paho-mqtt==1.6.1

# I2C Communication (Linux only)
smbus2==0.4.3; sys_platform == 'linux'

# Serial Communication (for PMS5003)
pyserial==3.5

# GPIO Control (Raspberry Pi only)
RPi.GPIO==0.7.1; sys_platform == 'linux'

# Configuration
PyYAML==6.0.1

# Utilities
python-dateutil==2.8.2

# mDNS Service Discovery
zeroconf==0.132.2

# Sensor-specific libraries (Linux only)
bme680==1.1.1; sys_platform == 'linux'
adafruit-circuitpython-scd4x==1.3.8; sys_platform == 'linux'
adafruit-circuitpython-bh1750==1.1.7; sys_platform == 'linux'
